util: Move predicates.go to util package
