+++ controllers/util/predicates.go
@@ -1,28 +1,28 @@
 // SPDX-FileCopyrightText: The RamenDR authors
 // SPDX-License-Identifier: Apache-2.0
 
-package controllers
+package util
 
 import (
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 )
 
-type createOrDeleteOrResourceVersionUpdatePredicate struct{}
+type CreateOrDeleteOrResourceVersionUpdatePredicate struct{}
 
-func (createOrDeleteOrResourceVersionUpdatePredicate) Create(e event.CreateEvent) bool {
+func (CreateOrDeleteOrResourceVersionUpdatePredicate) Create(e event.CreateEvent) bool {
 	return true
 }
 
-func (createOrDeleteOrResourceVersionUpdatePredicate) Delete(e event.DeleteEvent) bool {
+func (CreateOrDeleteOrResourceVersionUpdatePredicate) Delete(e event.DeleteEvent) bool {
 	return true
 }
 
-func (createOrDeleteOrResourceVersionUpdatePredicate) Update(e event.UpdateEvent) bool {
+func (CreateOrDeleteOrResourceVersionUpdatePredicate) Update(e event.UpdateEvent) bool {
 	return predicate.ResourceVersionChangedPredicate{}.Update(e)
 }
 
-func (createOrDeleteOrResourceVersionUpdatePredicate) Generic(e event.GenericEvent) bool {
+func (CreateOrDeleteOrResourceVersionUpdatePredicate) Generic(e event.GenericEvent) bool {
 	return false
 }
