+++ controllers/vrg_kubeobjects.go
@@ -709,7 +709,7 @@ func kubeObjectsRequestsWatch(b *builder.Builder, kubeObjects kubeobjects.Reques
 
 				return []reconcile.Request{}
 			}),
-			builder.WithPredicates(ResourceVersionUpdatePredicate{}),
+			builder.WithPredicates(util.ResourceVersionUpdatePredicate{}),
 		)
 	}
